+++ MapboxSearch/ui/src/main/java/com/mapbox/search/ui/utils/extenstion/EditText.kt
@@ -1,7 +1,0 @@
-package com.mapbox.search.ui.utils.extenstion
-
-import android.widget.EditText
-
-internal fun EditText.setCursorAtTheEnd() {
-    setSelection(text.length)
-}